add `base/events` tests
