+++ src/base/events/index.ts
@@ -1,3 +1,6 @@
 export * from './DisposalBin';
-export * from './events';
 export * from './hostedEventListener';
+export * from './isPointerEvent';
+export * from './listen';
+export * from './redispatchEvent';
+export * from './VdsEvent';
