+++ src/base/events/isPointerEvent.ts
@@ -1,0 +1,5 @@
+export function isPointerEvent(
+  event: Event | undefined
+): event is PointerEvent {
+  return event?.type.includes('pointer') ?? false;
+}
