+++ src/base/events/tests/hostedEventListener.test.ts
@@ -1,0 +1,3 @@
+test('', () => {
+  //
+});
